+++ src/main/java/com/nick1est/proconnectx/dto/ServiceFilter.java
@@ -3,6 +3,7 @@ package com.nick1est.proconnectx.dto;
 import com.nick1est.proconnectx.dao.CategoryType;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -10,7 +11,7 @@ public class ServiceFilter {
     private String title;
     private List<CategoryType> categories;
     private String location;
-    private Double rating;
-    private Double minBudget;
-    private Double maxBudget;
+    private BigDecimal rating;
+    private BigDecimal minBudget;
+    private BigDecimal maxBudget;
 }
